v0.1.0

+++ src/repositories/app.repository.ts
@@ -1,30 +1,38 @@
 import Boom from '@hapi/boom';
-import App from '../models/app.model';
-import Build from '../models/build.model';
-import admin from 'firebase-admin';
+import { AppModel } from '../models/app.model';
+import { BuildModel } from '../models/build.model';
+import { AppDto } from '../dtos/AppDto';
+import { Bucket } from '@google-cloud/storage';
 
 export interface AppRepository {
-  getApp: (appId: number) => Promise<object>;
-  deleteApp: (appId: number) => Promise<any>;
+  getApp: (appId: number) => Promise<AppDto>;
+  deleteApp: (appId: number) => Promise<void>;
 }
 
-const appRepository = (): AppRepository => {
-  const bucket = admin.storage().bucket();
+const appRepository = (
+  bucket: Bucket,
+  App: AppModel,
+  Build: BuildModel
+): AppRepository => {
   return {
-    getApp: async appId => {
+    getApp: async (appId) => {
       const app = await App.findByPk(appId, {
         include: [{ model: Build, as: 'builds' }],
       });
 
       if (!app) throw Boom.notFound();
-      return app.toJSON();
+
+      return <AppDto>app.toJSON();
     },
-    deleteApp: async appId => {
+    deleteApp: async (appId) => {
       const app = await App.findByPk(appId);
+      if (!app) throw Boom.notFound();
+
       await bucket.deleteFiles({
-        prefix: 'bundles/' + app?.bundleIdentifier + '/',
+        prefix: 'bundles/' + app.bundleIdentifier + '/',
       });
-      return app?.destroy();
+
+      return app.destroy();
     },
   };
 };
